Update project3.cpp
- Added variable long long int to increase the results over 1.000.000.000

+++ projects/project3.cpp
@@ -2,21 +2,22 @@
 
 int main() {
 
-	int numeroBase, numeroMax, expoente, i;
+	long long int numeroBase, numeroMax, expoente, i;
 	
 	printf("Qual o número da base?\n");
-	scanf("%d", &numeroBase);
+	scanf("%lld", &numeroBase);
 	printf("Qual o número da potência máxima?\n");
-	scanf("%d", &numeroMax);
+	scanf("%lld", &numeroMax);
 	//input: A base number and a maximum number to iterate in the power of that number
 	expoente = numeroBase;
 	
 	for(i = 1; i <= numeroMax; i++) {
 		// The variable "i" here is used just to repeat the power process and is not directly relationed to the operation
 		numeroBase = numeroBase * expoente;
-		printf("\n%d", numeroBase);
+		printf("\n%lld", numeroBase);
 	}
 	//output: The power of the chosed base time the maximum number 
 	return 0;
 }
+
